cleaned up stages models
