+++ models/Stages/updateStage.ts
@@ -1,0 +1,45 @@
+import { UpdateCommand } from "@aws-sdk/lib-dynamodb";
+import { Dynamo } from "../../awsClients/ddbDocClient";
+import { FORBIDDEN_PROPERTIES, ENTITY_TYPES } from "../../Config";
+import { UpdateStageInput } from "../../types/main";
+const { DYNAMO_TABLE_NAME } = process.env;
+
+export default async function update(props: UpdateStageInput): Promise<void> {
+  const { orgId, stageId, newStageValues } = props;
+
+  // Build update expression
+  let allUpdateExpressions: string[] = [];
+  let allAttributeValues: any = {};
+
+  // Filter out forbidden property
+  for (const property in newStageValues) {
+    if (FORBIDDEN_PROPERTIES.STAGE.includes(property)) {
+      // Delete the property so it isn't updated
+      delete newStageValues[property];
+    }
+
+    // If its a valid property, start creating the new update expression
+    // Push each property into the update expression
+    allUpdateExpressions.push(`${property} = :${property}`);
+
+    // Create values for each attribute
+    allAttributeValues[`:${property}`] = newStageValues[property];
+  }
+
+  const params = {
+    Key: {
+      PK: `${ENTITY_TYPES.ORG}#${orgId}#${ENTITY_TYPES.STAGE}#${stageId}`,
+      SK: ENTITY_TYPES.STAGE,
+    },
+    UpdateExpression: `SET ` + allUpdateExpressions.join(", "),
+    ExpressionAttributeValues: allAttributeValues,
+    TableName: DYNAMO_TABLE_NAME,
+    ConditionExpression: "attribute_exists(PK)",
+  };
+
+  try {
+    await Dynamo.send(new UpdateCommand(params));
+  } catch (error) {
+    throw new Error(error);
+  }
+}
